Refactoring: Used the function `normalize_data_format` to remove some code. (#10690)
* Used the function `normalize_data_format` to remove some code. Moved the implementation to keras/backend/common.py.

* Changed all conv_utils.normalize_data_format into K.normalize_data_format
