+++ keras/utils/conv_utils.py
@@ -48,17 +48,6 @@ def normalize_tuple(value, n, name):
     return value_tuple
 
 
-def normalize_data_format(value):
-    if value is None:
-        value = K.image_data_format()
-    data_format = value.lower()
-    if data_format not in {'channels_first', 'channels_last'}:
-        raise ValueError('The `data_format` argument must be one of '
-                         '"channels_first", "channels_last". Received: ' +
-                         str(value))
-    return data_format
-
-
 def normalize_padding(value):
     padding = value.lower()
     allowed = {'valid', 'same', 'causal'}
